requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asn1crypto==1.5.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.0
click==8.2.1
cryptography==44.0.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
Faker==33.1.0
fastapi==0.116.1
filelock==3.16.1
Flask==3.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
passlib==1.7.4
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.3
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.43
starlette==0.47.3
tomlkit==0.13.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
